feat: animated bg in ui package
