+++ packages/ui/src/base.css
@@ -69,18 +69,4 @@
   input[type="number"] {
     -moz-appearance: textfield;
   }
-
-  .bg-grid-pattern {
-    @apply bg-transparent bg-repeat opacity-90;
-
-    background-image: linear-gradient(#fff 1px, transparent 1px),
-      linear-gradient(to right, #fff 1px, transparent 1px);
-    background-size: 20px 20px;
-  }
-
-  html[data-theme="dark"] .bg-grid-pattern {
-    background-image: linear-gradient(#222 1px, transparent 1px),
-      linear-gradient(to right, #222 1px, transparent 1px);
-    background-size: 20px 20px;
-  }
 }
