+++ packages/ui/src/components/AnimatedBackground/AnimatedBackground.spec.tsx
@@ -1,11 +1,11 @@
 import { composeStories } from "@storybook/react";
 import { render } from "@testing-library/react";
 
-import * as stories from "./AnimatedBlobBackground.stories";
+import * as stories from "./AnimatedBackground.stories";
 
 const { Default } = composeStories(stories);
 
-describe("AnimatedBlobBackground", () => {
+describe("AnimatedBackground", () => {
   it("renders Default component story without breaking", () => {
     const { container } = render(<Default />);
     expect(container).toBeVisible();
