+++ packages/ui/src/components/AnimatedBackground/index.ts
@@ -1,0 +1,1 @@
+export * from "./AnimatedBackground";
