--- packages/ui/src/components/AnimatedBlobBackground/index.ts
@@ -1,1 +1,0 @@
-export * from "./AnimatedBlobBackground";
