+++ packages/ui/src/components/index.ts
@@ -1,5 +1,5 @@
 export * from "./Alert";
-export * from "./AnimatedBlobBackground";
+export * from "./AnimatedBackground";
 export * from "./Badge";
 export * from "./Button";
 export * from "./Card";
